Fix TS circular dependencies.
