+++ package.json
@@ -17,7 +17,7 @@
     "dev:beta": "BETA=true npm run dev",
     "lint": "npm-run-all lint:*",
     "lint:js": "eslint src",
-    "circular": "madge --circular ./src",
+    "circular": "madge --circular ./src --extensions js,ts,tsx",
     "start": "NODE_ENV=development webpack --config config/webpack.config.js --watch  --mode=development",
     "start:beta": "BETA=true npm run start",
     "test": "jest --forceExit --detectOpenHandles",
