+++ src/chrome/create-chrome.test.js
@@ -1,10 +1,9 @@
-import createChromeInstance from './create-chrome';
+import { createChromeContext } from './create-chrome';
 
 jest.mock('../jwt/jwt');
 jest.mock('../auth/fetchPermissions');
 
 describe('create chrome', () => {
-  const insights = {};
   let jwt;
   beforeEach(() => {
     jwt = {
@@ -17,16 +16,22 @@ describe('create chrome', () => {
   });
 
   it('should create chrome instance', () => {
-    const { chrome } = createChromeInstance(jwt, insights);
+    const chrome = createChromeContext({
+      libJwt: jwt,
+      store: {},
+      getUser: () => Promise.resolve(),
+    });
     expect(chrome).toEqual(expect.any(Object));
   });
 
   it('should postpone getUserPermissions resolve, after chrome cache is initialized', () => {
     const promiseSpy = jest.fn();
     expect.assertions(1);
-    const {
-      chrome: { getUserPermissions },
-    } = createChromeInstance(jwt, insights);
+    const { getUserPermissions } = createChromeContext({
+      libJwt: jwt,
+      store: {},
+      getUser: () => Promise.resolve(),
+    });
     return getUserPermissions(promiseSpy).then(() => {
       expect(promiseSpy).toHaveBeenCalledWith('mocked-user-permissions');
     });
