+++ src/components/GlobalFilter/GlobalFilter.tsx
@@ -3,11 +3,12 @@ import PropTypes from 'prop-types';
 import { batch, shallowEqual, useDispatch, useSelector } from 'react-redux';
 import { useTagsFilter } from '@redhat-cloud-services/frontend-components/FilterHooks';
 import { fetchAllSIDs, fetchAllTags, fetchAllWorkloads, globalFilterChange } from '../../redux/actions';
+import { generateFilter } from './globalFilterApi';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { GlobalFilterDropdown, GlobalFilterDropdownProps } from './GlobalFilterMenu';
 import { storeFilter } from './filterApi';
 import { GlobalFilterTag, GlobalFilterWorkloads, ReduxState, SID } from '../../redux/store';
-import { FlagTagsFilter, generateFilter } from './globalFilterApi';
+import { FlagTagsFilter } from '../../@types/types';
 import { isGlobalFilterAllowed } from '../../utils/common';
 
 const useLoadTags = (hasAccess = false) => {
