+++ src/components/Navigation/ChromeNavGroup.tsx
@@ -1,32 +1,8 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import { NavGroup } from '@patternfly/react-core';
 
-import WrenchIcon from '@patternfly/react-icons/dist/js/icons/wrench-icon';
-import SecurityIcon from '@patternfly/react-icons/dist/js/icons/security-icon';
-import TrendUpIcon from '@patternfly/react-icons/dist/js/icons/trend-up-icon';
-import CodeIcon from '@patternfly/react-icons/dist/js/icons/code-icon';
-import DatabaseIcon from '@patternfly/react-icons/dist/js/icons/database-icon';
-import CloudIcon from '@patternfly/react-icons/dist/js/icons/cloud-upload-alt-icon';
-
-import ChromeNavItemFactory from './ChromeNavItemFactory';
-import { NavItem } from '../../@types/types';
-
-const sectionTitleMapper = {
-  wrench: <WrenchIcon />,
-  shield: <SecurityIcon />,
-  database: <DatabaseIcon />,
-  cloud: <CloudIcon />,
-  code: <CodeIcon />,
-  'trend-up': <TrendUpIcon />,
-};
-
-export type ChromeNavGroupProps = {
-  navItems: NavItem[];
-  isHidden?: boolean;
-  icon?: keyof typeof sectionTitleMapper;
-  title: string;
-};
+import ChromeNavItemFactory, { sectionTitleMapper } from './ChromeNavItemFactory';
+import { ChromeNavGroupProps } from '../../@types/types';
 
 const ChromeNavGroup = ({ navItems, isHidden, icon, title }: ChromeNavGroupProps) => {
   const filteredFedrampNavItems = navItems;
@@ -49,13 +25,6 @@ const ChromeNavGroup = ({ navItems, isHidden, icon, title }: ChromeNavGroupProps
       ))}
     </NavGroup>
   );
-};
-
-ChromeNavGroup.propTypes = {
-  navItems: PropTypes.array.isRequired,
-  icon: PropTypes.oneOf(['wrench', 'shield', 'trend-up', 'database', 'cloud', 'code']),
-  title: PropTypes.string.isRequired,
-  isHidden: PropTypes.bool,
 };
 
 export default ChromeNavGroup;
