+++ src/components/Navigation/navContext.ts
@@ -1,5 +1,5 @@
 import React, { createContext } from 'react';
-import componentMapper from './componentMapper';
+import { ChromeNavExapandableProps, ChromeNavGroupProps, ChromeNavItemProps, DynamicNavProps } from '../../@types/types';
 
 export type OnLinkClick = (
   event: React.MouseEvent<HTMLAnchorElement, MouseEvent> | React.MouseEvent<HTMLAnchorElement, MouseEvent>,
@@ -10,7 +10,12 @@ interface NavContext {
   onLinkClick?: OnLinkClick;
   isNavOpen?: boolean;
   inPageLayout?: boolean;
-  componentMapper: typeof componentMapper;
+  componentMapper: {
+    group: React.FC<ChromeNavGroupProps>;
+    expandable: React.FC<ChromeNavExapandableProps>;
+    item: React.FC<ChromeNavItemProps>;
+    dynamicNav: React.FC<DynamicNavProps>;
+  };
 }
 
 const NavContext = createContext<NavContext>({} as NavContext);
