+++ src/redux/chromeReducers.ts
@@ -1,10 +1,10 @@
 import { QuickStart } from '@patternfly/quickstarts';
 import { ChromeUser } from '@redhat-cloud-services/types';
 import { REQUESTS_COUNT, REQUESTS_DATA } from '../utils/consts';
-import { NavItem } from '../@types/types';
+import { ChromeModule, NavItem, Navigation } from '../@types/types';
 import { generateRoutesList, highlightItems, isBeta, isFedRamp, levelArray } from '../utils/common';
 import { ThreeScaleError } from '../utils/responseInterceptors';
-import { AccessRequest, ChromeModule, ChromeState, Navigation } from './store';
+import { AccessRequest, ChromeState } from './store';
 
 export function contextSwitcherBannerReducer(state: ChromeState): ChromeState {
   return {
