+++ src/utils/useRenderFedramp.ts
@@ -1,8 +1,9 @@
 import { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { matchPath } from 'react-router-dom';
+import { ChromeModule } from '../@types/types';
 
-import { ChromeModule, ReduxState } from '../redux/store';
+import { ReduxState } from '../redux/store';
 import { isFedRamp } from './common';
 const isFedrampEnv = isFedRamp();
